% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCApp-package.R
\docType{package}
\name{PCApp-package}
\alias{PCApp}
\alias{PCApp-package}
\title{PCApp: PCApp}
\description{
PCApp_DESC.
}
\author{
\strong{Maintainer}: Justus Weyers \email{AUTHOR@MAIL.COM}

}
\keyword{internal}
